Refactor: Consolidate knowledge-base into knowledge-service
This commit reorganizes the project structure by merging the `knowledge-base` directory into the `knowledge-service` directory. This consolidation aims to streamline the codebase and improve maintainability.

Additionally, the commit updates database models to use `BigInteger` for IDs, ensuring better scalability for large datasets. The `prepare_data.py` script is also adjusted to reflect the new path for knowledge base data.

+++ backend/agent-service/app/common/models/document.py
@@ -5,7 +5,7 @@ Document Model
 SQLAlchemy model for Document entity.
 """
 
-from sqlalchemy import JSON, Column, ForeignKey, Integer, String, Text
+from sqlalchemy import JSON, BigInteger, Column, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import relationship
 
 from .base import BaseModel
@@ -17,7 +17,7 @@ class Document(BaseModel):
     __tablename__ = "documents"
 
     project_id = Column(
-        Integer,
+        BigInteger,
         ForeignKey("projects.id", ondelete="CASCADE"),
         nullable=False,
         index=True,
